fix(fdr2csv): refactor to unify fdr2csv, fix conversion bug (#9324)
* feat(fdr2csv): refactor to unify fdr2csv, fix conversion bug

* chore: don't downgrade crate versions

* feat(fdr2csv): add option to get raw file version

* fix(fdr): fixed wrong storage of parameters

* feat(fdr2csv): -o no longer required when -r or -g are present

* fix(fdr2csv): correct description of parameters

---------
