+++ tools/fdr2csv/a320_wrapper.hpp
@@ -1,0 +1,7 @@
+#include "../../fbw-a32nx/src/wasm/fbw_a320/src/model/AutopilotLaws_types.h"
+#include "../../fbw-a32nx/src/wasm/fbw_a320/src/model/AutopilotStateMachine_types.h"
+#include "../../fbw-a32nx/src/wasm/fbw_a320/src/model/Autothrust_types.h"
+#include "../../fbw-a32nx/src/wasm/fbw_a320/src/model/ElacComputer_types.h"
+#include "../../fbw-a32nx/src/wasm/fbw_a320/src/model/FacComputer_types.h"
+#include "../../fbw-a32nx/src/wasm/fbw_a320/src/model/SecComputer_types.h"
+#include "../../fbw-a32nx/src/wasm/fbw_a320/src/recording/RecordingDataTypes.h"
