+++ tools/fdr2csv/a380_wrapper.hpp
@@ -1,0 +1,8 @@
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/interface/FuelSystemData.h"
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/model/A380PrimComputer_types.h"
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/model/A380SecComputer_types.h"
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/model/AutopilotLaws_types.h"
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/model/AutopilotStateMachine_types.h"
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/model/Autothrust_types.h"
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/model/FacComputer_types.h"
+#include "../../fbw-a380x/src/wasm/fbw_a380/src/recording/RecordingDataTypes.h"
