--- tools/fdr2csv/common/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "fdr2csv_common"
-version = "0.1.0"
-authors = ["FlyByWire Simulations"]
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[build-dependencies]
-bindgen = "0.69.5"
-
-[dependencies]
-clap = { version = "4.5.20", features = ["derive"] }
-serde = { version = "1.0.213", features = ["derive"] }
-bytemuck = { version = "1.19.0", features = ["derive"] }
-
-csv = "1.3.0"
-flate2 = "1.0.34"
-
-[workspace]
