--- tools/fdr2csv/common/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod csv_header_serializer;
-pub mod error;
