+++ tools/fdr2csv/src/a320_headers.rs
@@ -7,4 +7,4 @@ use serde::Serialize;
 
 use bytemuck::AnyBitPattern;
 
-include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
+include!(concat!(env!("OUT_DIR"), "/bindings_320.rs"));
