+++ tools/fdr2csv/src/a380_headers.rs
@@ -7,4 +7,4 @@ use serde::Serialize;
 
 use bytemuck::AnyBitPattern;
 
-include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
+include!(concat!(env!("OUT_DIR"), "/bindings_380.rs"));
