+++ tools/fdr2csv/src/error.rs
